refactor search example
