+++ examples/flutter_main/lib/counter/counter.dart
@@ -1,0 +1,7 @@
+import 'package:example/const.dart';
+import 'package:flutter/material.dart';
+import 'package:provider/provider.dart';
+import 'package:state_beacon/state_beacon.dart';
+
+part 'view.dart';
+part 'controller.dart';
