+++ examples/flutter_main/lib/main.dart
@@ -1,9 +1,9 @@
 // ignore_for_file: prefer_const_constructors, prefer_const_literals_to_create_immutables
 
-import 'package:example/counter_page.dart';
+import 'package:example/counter/counter.dart';
 import 'package:example/infinite_list/infinite_list.dart';
 import 'package:example/konami_page.dart';
-import 'package:example/search_page.dart';
+import 'package:example/search/search.dart';
 import 'package:example/todo_page.dart';
 import 'package:flutter/material.dart';
 import 'package:state_beacon/state_beacon.dart';
