+++ examples/flutter_main/lib/search/models.dart
@@ -1,3 +1,5 @@
+part of 'search.dart';
+
 class Weather {
   final String cityName;
   final double temperatureCelsius;
